+++ packages/terra-application-layout/src/terra-dev-site/doc/example/ExampleApplicationWithNavIcons.jsx
@@ -218,7 +218,7 @@ const userData = {
  */
 const nameConfig = Object.freeze({
   title: 'Example Application',
-  accessory: <Image variant="rounded" src="https://github.com/cerner/terra-framework/raw/master/terra.png" height="26px" width="26px" />,
+  accessory: <Image variant="rounded" src="https://github.com/cerner/terra-framework/raw/master/terra.png" alt="App logo" height="26px" width="26px" />,
 });
 
 class ExampleApplication extends React.Component {