+++ packages/terra-embedded-content-consumer/src/terra-dev-site/test/embedded-content-consumer/consumers/BasicConsumer.test.jsx
@@ -1,4 +1,3 @@
-/* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
 import { Consumer } from 'xfc';
 import EmbeddedContentConsumer from '../../../../EmbeddedContentConsumer';
@@ -7,7 +6,7 @@ Consumer.init();
 
 const BasicConsumer = () => (
   <EmbeddedContentConsumer
-    src="#/raw/tests/terra-embedded-content-consumer/embedded-content-consumer/providers/basic-provider"
+    src="#/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/basic-provider"
   />
 );