+++ packages/terra-date-picker/src/DatePicker.jsx
@@ -5,7 +5,6 @@ import ResponsiveElement from 'terra-responsive-element';
 
 /* eslint-disable-next-line  */
 import ReactDatePicker from './react-datepicker';
-import PopperContainer from './_PopperContainer';
 import DateInput from './DateInput';
 import DateUtil from './DateUtil';
 import styles from './DatePicker.module.scss';
@@ -300,7 +299,7 @@ class DatePicker extends React.Component {
     const maxMomentDate = DateUtil.createSafeDate(maxDate);
     const minMomentDate = DateUtil.createSafeDate(minDate);
 
-    const portalPicker = (
+    const hookshotPicker = (
       <ReactDatePicker
         {...customProps}
         selected={this.state.selectedDate}
@@ -373,15 +372,6 @@ class DatePicker extends React.Component {
         fixedHeight
         locale={intl.locale}
         placeholderText={placeholderText}
-        popperContainer={PopperContainer}
-        popperPlacement="bottom"
-        popperModifiers={{
-          keepTogether: {
-            enabled: true,
-          },
-        }}
-        popoverAttachment="top center"
-        popoverTargetAttachment="bottom center"
         dropdownMode="select"
         showMonthDropdown
         showYearDropdown
@@ -397,7 +387,7 @@ class DatePicker extends React.Component {
       >
         <ResponsiveElement
           responsiveTo="window"
-          defaultElement={portalPicker}
+          defaultElement={hookshotPicker}
           small={popupPicker}
         />
       </div>