+++ packages/terra-embedded-content-consumer/CHANGELOG.md
@@ -3,6 +3,15 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* EmbeddedContentProviderTestTemplate for testing embedded applications
+
+### Changed
+* Updated existing provider examples to use the EmbeddedContentProviderTestTemplate
+* Changed Provider examples to use `.provider` extension to fix provider examples in terra-ui
+
+### Removed
+* Dead code tests / examples that tested XFC implementation
 
 3.2.0 - (March 26, 2019)
 ------------------